Adds notification on main queue
This is part of the crux of the fix. Since the adding of animations ot the layer is on the main queue, the resetting of an animation and setting its time can easily happen BEFORE the animations are ready on the layer.

So, we need a way of tracking when all the animations have been added properly, for each instance of Animation in a timeline.

Sometimes, we may NOT want to make this announcement, so we add a notify parameter to a couple of methods as well.

We also add callback methods for when the animation is complete, including a protocol for AnimationDelegate that we can use to track when an instance is ready to be manipulated.