Track resetting of animations
Resetting animations triggers the asynchronous removal and addition of CAKeyframe animations to a given layer. Since this process is async, we need to track when all of the animations have been reset.

Adding a dispatch group allows us to track this properly.

Setting all animations to have the timeline as a delegate is required.