Handle autoplaying an animation
In certain circumstances, we may want to immediately stop, switch timelines, adjust the time of the new timeline, and trigger playback. In the case of a toggle button this will allow the animation to stop in place and appear to "play backwards" from any arbitrary time – instead of resetting to the end of the animnation and then initiating playback.

In many instances, calling 'play()' will happen before the asynchronous reset of a Timeline's animations is complete. So, we need to track whether or not the timeline should autoplay.