Fix setting of animation time
Now, for the actual fix. I found a bug in swapping and resetting the time of an animation. The fix for which involves a bit of the timeline, and of the animation class.

First, we fix the language in offset to avoid misunderstanding and conflict with the `time` variable.

We remove the delegate call because if we do this here, it will actually be called before the timeline is actually reset (because of the async issues we fixed in the two previous commits.