Properly offset an animation
This requires setting the current time offset of each animation, rather than just the layer's time offset. It also requires setting the begintime of the layer.

This also requires a slightly different technique for handling the timeOffset – we need to set this on each keyframeAnimation when hitting `play()`.

We also move the `layer.speed = 0` to the pause method.