Improve control over resetting and setting the current time of Timelines (#24)
* Adds notification on main queue

This is part of the crux of the fix. Since the adding of animations ot the layer is on the main queue, the resetting of an animation and setting its time can easily happen BEFORE the animations are ready on the layer.

So, we need a way of tracking when all the animations have been added properly, for each instance of Animation in a timeline.

Sometimes, we may NOT want to make this announcement, so we add a notify parameter to a couple of methods as well.

We also add callback methods for when the animation is complete, including a protocol for AnimationDelegate that we can use to track when an instance is ready to be manipulated.

* Track resetting of animations

Resetting animations triggers the asynchronous removal and addition of CAKeyframe animations to a given layer. Since this process is async, we need to track when all of the animations have been reset.

Adding a dispatch group allows us to track this properly.

Setting all animations to have the timeline as a delegate is required.

* Handle autoplaying an animation

In certain circumstances, we may want to immediately stop, switch timelines, adjust the time of the new timeline, and trigger playback. In the case of a toggle button this will allow the animation to stop in place and appear to "play backwards" from any arbitrary time – instead of resetting to the end of the animnation and then initiating playback.

In many instances, calling 'play()' will happen before the asynchronous reset of a Timeline's animations is complete. So, we need to track whether or not the timeline should autoplay.

* Fix setting of animation time

Now, for the actual fix. I found a bug in swapping and resetting the time of an animation. The fix for which involves a bit of the timeline, and of the animation class.

First, we fix the language in offset to avoid misunderstanding and conflict with the `time` variable.

We remove the delegate call because if we do this here, it will actually be called before the timeline is actually reset (because of the async issues we fixed in the two previous commits.

* Properly offset an animation

This requires setting the current time offset of each animation, rather than just the layer's time offset. It also requires setting the begintime of the layer.

This also requires a slightly different technique for handling the timeOffset – we need to set this on each keyframeAnimation when hitting `play()`.

We also move the `layer.speed = 0` to the pause method.

* Fix language, avoid misinterpretation

* Linting

* code review update

Co-authored-by: Jeremie Benhamron <jeremiebenhamron@gmail.com>